requirements.txt
pandas
requests
openai
ruff
bs4
markdown
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
apache-airflow
moviepy